+++ extensions/ql-vscode/src/codeql-cli/cli-version.ts
@@ -10,9 +10,7 @@ interface VersionResult {
 }
 
 export interface CliFeatures {
-  featuresInVersionResult?: boolean;
-  mrvaPackCreate?: boolean;
-  generateSummarySymbolMap?: boolean;
+  queryServerRunQueries?: boolean;
 }
 
 export interface VersionAndFeatures {