+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -23,7 +23,8 @@ import { URLSearchParams } from "url";
 import { DisposableObject } from "../common/disposable-object";
 import { ONE_HOUR_IN_MS, TWO_HOURS_IN_MS } from "../common/time";
 import { assertNever, getErrorMessage } from "../common/helpers-pure";
-import type { CompletedLocalQueryInfo, LocalQueryInfo } from "../query-results";
+import type { CompletedLocalQueryInfo } from "../query-results";
+import { LocalQueryInfo } from "../query-results";
 import type { QueryHistoryInfo } from "./query-history-info";
 import {
   getActionsWorkflowRunUrl,
@@ -50,7 +51,10 @@ import type { QueryRunner } from "../query-server";
 import type { VariantAnalysisManager } from "../variant-analysis/variant-analysis-manager";
 import type { VariantAnalysisHistoryItem } from "./variant-analysis-history-item";
 import { getTotalResultCount } from "../variant-analysis/shared/variant-analysis";
-import { HistoryTreeDataProvider } from "./history-tree-data-provider";
+import {
+  HistoryTreeDataProvider,
+  SortOrder,
+} from "./history-tree-data-provider";
 import type { QueryHistoryDirs } from "./query-history-dirs";
 import type { QueryHistoryCommands } from "../common/commands";
 import type { App } from "../common/app";
@@ -96,15 +100,6 @@ const SHOW_QUERY_TEXT_QUICK_EVAL_MSG = `\
 ////////////////////////////////////////////////////////////////////////////////////
 
 `;
-
-enum SortOrder {
-  NameAsc = "NameAsc",
-  NameDesc = "NameDesc",
-  DateAsc = "DateAsc",
-  DateDesc = "DateDesc",
-  CountAsc = "CountAsc",
-  CountDesc = "CountDesc",
-}
 
 /**
  * Number of milliseconds two clicks have to arrive apart to be
@@ -337,6 +332,11 @@ export class QueryHistoryManager extends DisposableObject {
         this.handleOpenOnGithub.bind(this),
         "query",
       ),
+      "codeQLQueryHistory.viewAutofixes": createSingleSelectionCommand(
+        this.app.logger,
+        this.handleViewAutofixes.bind(this),
+        "query",
+      ),
       "codeQLQueryHistory.copyRepoList": createSingleSelectionCommand(
         this.app.logger,
         this.handleCopyRepoList.bind(this),
@@ -348,8 +348,37 @@ export class QueryHistoryManager extends DisposableObject {
     };
   }
 
-  public completeQuery(info: LocalQueryInfo, results: QueryWithResults): void {
-    info.completeThisQuery(results);
+  public completeQueries(
+    info: LocalQueryInfo,
+    results: QueryWithResults[],
+  ): void {
+    let first = true;
+    // Sorting results by the output/results basename should produce a deterministic order.
+    results.sort((a, b) => {
+      const aPath = a.query.outputBaseName;
+      const bPath = b.query.outputBaseName;
+      return aPath.localeCompare(bPath);
+    });
+    for (const result of results) {
+      if (first) {
+        // This is the first query, so we can just update the existing info.
+        info.completeThisQuery(result);
+        first = false;
+      } else {
+        // For other queries in the same run, we'll add new entries to the history pane. In the long
+        // term, it would be better if we could have a single entry containing sub-entries for each
+        // query.
+        const clonedInfo = new LocalQueryInfo(
+          info.initialInfo,
+          undefined,
+          info.failureReason,
+          undefined,
+          info.evaluatorLogPaths,
+        );
+        clonedInfo.completeThisQuery(result);
+        this.addQuery(clonedInfo);
+      }
+    }
     this._onDidCompleteQuery.fire(info);
   }
 
@@ -551,6 +580,23 @@ export class QueryHistoryManager extends DisposableObject {
           await this.removeVariantAnalysis(item);
         } else {
           assertNever(item);
+        }
+      }),
+    );
+
+    await Promise.all(
+      this.treeDataProvider.allHistory.map(async (item) => {
+        // Remove any local queries whose directories no longer exist. This can happen when running
+        // a query suite, which produces multiple queries in the history pane that all share the
+        // same underlying directory, which we may have just deleted above. (Ideally, there would be
+        // a first-class concept of a local multi-query run in this pane that would group them all
+        // together, but doing it this way at least avoids cluttering the history pane with entries
+        // that can no longer be viewed).
+        if (item.t === "local") {
+          const dir = item.completedQuery?.query.querySaveDir;
+          if (dir && !(await pathExists(dir))) {
+            this.treeDataProvider.remove(item);
+          }
         }
       }),
     );
@@ -942,7 +988,7 @@ export class QueryHistoryManager extends DisposableObject {
     if (hasInterpretedResults) {
       await tryOpenExternalFile(
         this.app.commands,
-        query.resultsPaths.interpretedResultsPath,
+        query.interpretedResultsPath,
       );
     } else {
       const label = this.labelProvider.getLabel(item);
@@ -1003,6 +1049,14 @@ export class QueryHistoryManager extends DisposableObject {
       "vscode.open",
       Uri.parse(actionsWorkflowRunUrl),
     );
+  }
+
+  async handleViewAutofixes(item: QueryHistoryInfo) {
+    if (item.t !== "variant-analysis") {
+      return;
+    }
+
+    await this.variantAnalysisManager.viewAutofixes(item.variantAnalysis.id);
   }
 
   async handleCopyRepoList(item: QueryHistoryInfo) {