+++ extensions/ql-vscode/src/run-queries-shared.ts
@@ -27,7 +27,6 @@ import type {
 import type { BaseLogger } from "./common/logging";
 import { showAndLogWarningMessage } from "./common/logging";
 import { extLogger } from "./common/logging/vscode";
-import { generateSummarySymbolsFile } from "./log-insights/summary-parser";
 import { getErrorMessage } from "./common/helpers-pure";
 import { createHash } from "crypto";
 import { QueryOutputDir } from "./local-queries/query-output-dir";
@@ -65,6 +64,7 @@ export class QueryEvaluationInfo extends QueryOutputDir {
    */
   constructor(
     querySaveDir: string,
+    public readonly outputBaseName: string,
     public readonly dbItemPath: string,
     public readonly databaseHasMetadataFile: boolean,
     public readonly quickEvalPosition?: Position,
@@ -73,23 +73,30 @@ export class QueryEvaluationInfo extends QueryOutputDir {
     super(querySaveDir);
   }
 
-  get resultsPaths() {
-    return {
-      resultsPath: this.bqrsPath,
-      interpretedResultsPath: join(
-        this.querySaveDir,
-        this.metadata?.kind === "graph"
-          ? "graphResults"
-          : "interpretedResults.sarif",
-      ),
-    };
+  get resultsPath() {
+    return this.getBqrsPath(this.outputBaseName);
   }
+
+  get interpretedResultsPath() {
+    return this.getInterpretedResultsPath(
+      this.metadata?.kind,
+      this.outputBaseName,
+    );
+  }
+
+  get csvPath() {
+    return this.getCsvPath(this.outputBaseName);
+  }
+
+  get dilPath() {
+    return this.getDilPath(this.outputBaseName);
+  }
+
   getSortedResultSetPath(resultSetName: string) {
     const hasher = createHash("sha256");
     hasher.update(resultSetName);
-    return join(
-      this.querySaveDir,
-      `sortedResults-${hasher.digest("hex")}.bqrs`,
+    return this.getBqrsPath(
+      `${this.outputBaseName}-sorted-${hasher.digest("hex")}`,
     );
   }
 
@@ -127,7 +134,7 @@ export class QueryEvaluationInfo extends QueryOutputDir {
    * Holds if this query actually has produced interpreted results.
    */
   async hasInterpretedResults(): Promise<boolean> {
-    return pathExists(this.resultsPaths.interpretedResultsPath);
+    return pathExists(this.interpretedResultsPath);
   }
 
   /**
@@ -205,7 +212,7 @@ export class QueryEvaluationInfo extends QueryOutputDir {
     let nextOffset: number | undefined = 0;
     do {
       const chunk: DecodedBqrsChunk = await cliServer.bqrsDecode(
-        this.resultsPaths.resultsPath,
+        this.resultsPath,
         resultSet,
         {
           pageSize: 100,
@@ -243,9 +250,9 @@ export class QueryEvaluationInfo extends QueryOutputDir {
    * If the query has no result sets, then return undefined.
    */
   async chooseResultSet(cliServer: CodeQLCliServer) {
-    const resultSets = (
-      await cliServer.bqrsInfo(this.resultsPaths.resultsPath)
-    )["result-sets"];
+    const resultSets = (await cliServer.bqrsInfo(this.resultsPath))[
+      "result-sets"
+    ];
     if (!resultSets.length) {
       return undefined;
     }
@@ -284,7 +291,7 @@ export class QueryEvaluationInfo extends QueryOutputDir {
     }
     await cliServer.generateResultsCsv(
       ensureMetadataIsComplete(this.metadata),
-      this.resultsPaths.resultsPath,
+      this.resultsPath,
       this.csvPath,
       sourceInfo,
     );
@@ -346,6 +353,23 @@ export function validateQueryPath(
       );
     }
   }
+}
+
+/**
+ * Validates that the specified URI represents a QL query suite (QLS), and returns the file system
+ * path to that suite.
+ */
+export function validateQuerySuiteUri(suiteUri: Uri): string {
+  if (suiteUri.scheme !== "file") {
+    throw new Error("Can only run queries that are on disk.");
+  }
+  const suitePath = suiteUri.fsPath;
+  if (!suitePath.endsWith(".qls")) {
+    throw new Error(
+      'The selected resource is not a CodeQL query suite; It should have the extension ".qls".',
+    );
+  }
+  return suitePath;
 }
 
 export interface QuickEvalContext {
@@ -545,15 +569,6 @@ export async function generateEvalLogSummaries(
 
     if (humanReadableSummary !== undefined) {
       summarySymbols = outputDir.evalLogSummarySymbolsPath;
-      if (
-        !(await cliServer.cliConstraints.supportsGenerateSummarySymbolMap())
-      ) {
-        // We're using an old CLI that cannot generate the summary symbols file while generating the
-        // human-readable log summary. As a fallback, create it by parsing the human-readable
-        // summary.
-        progress(progressUpdate(3, 3, "Generating summary symbols file"));
-        await generateSummarySymbolsFile(humanReadableSummary, summarySymbols);
-      }
     }
   }