+++ extensions/ql-vscode/src/variant-analysis/variant-analysis-manager.ts
@@ -73,6 +73,7 @@ import type {
   VariantAnalysisCommands,
 } from "../common/commands";
 import { exportVariantAnalysisResults } from "./export-results";
+import { viewAutofixesForVariantAnalysisResults } from "./view-autofixes";
 import {
   readRepoStates,
   REPO_STATES_FILENAME,
@@ -424,7 +425,7 @@ export class VariantAnalysisManager
         this.app.credentials,
         variantAnalysisSubmission,
       );
-    } catch (e: unknown) {
+    } catch (e) {
       // If the error is handled by the handleRequestError function, we don't need to throw
       if (
         e instanceof RequestError &&
@@ -964,6 +965,21 @@ export class VariantAnalysisManager
     await this.app.commands.execute(
       "vscode.open",
       Uri.parse(actionsWorkflowRunUrl),
+    );
+  }
+
+  public async viewAutofixes(
+    variantAnalysisId: number,
+    filterSort: RepositoriesFilterSortStateWithIds = defaultFilterSortState,
+  ) {
+    await viewAutofixesForVariantAnalysisResults(
+      this,
+      this.variantAnalysisResultsManager,
+      variantAnalysisId,
+      filterSort,
+      this.app.credentials,
+      this.app,
+      this.cliServer,
     );
   }